base

+++ src/base/Classes.hpp
@@ -740,7 +740,7 @@ public:
     const UnicodeString APrompt, UnicodeString &Value, const UnicodeString HelpKeyword,
     TStrings *History, bool PathInput,
     TInputDialogInitializeEvent OnInitialize, bool Echo) = 0;
-  virtual uintptr_t MoreMessageDialog(const UnicodeString AMessage,
+  virtual uint32_t MoreMessageDialog(const UnicodeString AMessage,
     TStrings *MoreMessages, TQueryType Type, uintptr_t Answers,
     const TMessageParams *Params) = 0;
 };
